diagnostics: add GetServiceStatus API and "status" tool
This change adds a new API to cros_healthd to check the status of the
service. Currently it returns if the external network health and
diagnostics mojo remotes have been bound in the service.

It also adds a new subtool to cros-health-tool, "status". This checks
and prints the status of the cros_healthd service.

BUG=b:178274447
TEST=FEATURES=test emerge-sarien diagnostics
(dut) cros-health-tool status (prints service status)

Change-Id: Ibccc9a2f99b45975da8be12b56cf6537c342acc8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2689559
Tested-by: Trent Begin <tbegin@chromium.org>
Commit-Queue: Trent Begin <tbegin@chromium.org>
Reviewed-by: Paul Moy <pmoy@chromium.org>