mount-encrypted: check the key decryption result size
This would prevent dm-crypt failed with Invalid argument error.

The original encryption key and system key are both 32 bytes.
Because we are using standard padding mode, The ciphertext would became
48 bytes(the encryption result of "32 bytes encryption key + '\x10' * 16")

The AesDecryptSpecifyBlockMode would only check the result have correct
padding. And which means the result may not have a correct length.

For example:
If the decryption result is "47 bytes + '\x01' * 1", it would
return a 47 bytes encryption key. And that isn't a correct result.
(The probability of this case is about 1/256)

Note: Standard Padding Pattern:
"32 bytes" + '\x10' * 16
...
"40 bytes" + '\x08' * 8
...
"47 bytes" + '\x01' * 1

Ref:
https://wiki.openssl.org/index.php/EVP_Symmetric_Encryption_and_Decryption#Padding

BUG=b:178679936, b:178332041
TEST=clear TPM && reboot && check mount-encrypted log

Change-Id: Idc22c0f6a878d81f9f91dce4a1c4837af0b54e21
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2690422
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Tested-by: Yi Chou <yich@google.com>
Commit-Queue: Yi Chou <yich@google.com>