shill: test-scripts: fixup configure-hidden-wifi
Dbus library needs a signature when providing a complex dictionary like
this, and we don't support introspection.

Also, shill rejects a Passphrase property for open networks, so only
provide it in the dict if it's legit.

And we expect to configure via 'SecurityClass' now, not 'Security'
(b/157935328).

BUG=none
TEST=`./configure-hidden-wifi nonexistent-ssid`

Change-Id: I0c1df93773d21dba3e66b519786fe003a121bfeb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2780059
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>