secanomalyd: Introduce secanomalyd.
This CL adds the basic skeleton of secanomalyd. secanomalyd is
(will be?) a security-anomaly detection daemon for Chrome OS, detecting
violations of Chrome OS security properties.

The first motivating security anomaly to report is the existence of
writable and executable mounts -- those should never occur in a
Chrome OS device in production. There is enough code in this CL to test
detection and reporting behaviour.

Several things still need to be done:
-Command-line flag to generate reports even in dev-mode.
-Unit tests.
-UMA metrics.
-crash_reporter-side implementation.

The daemon is currently disabled (there is no init file) and will not
be enabled until the above points are fixed.

See go/cros-reporting-suspicious-mounts for details.

BUG=chromium:1190150
TEST=Build, deploy, run manually, mounts are detected but not uploaded
(because test image.)

Change-Id: I513cc1c3b88131f1368b14a5ce9464ec44b1e501
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2773925
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>