typecd: Add a class to send notifications
Introduce a new class, NotificationManager, which can be used to send
relevant event notifications to other interested processes.

Start with a notification which informs the listener of a TBT device
being connected.

Currently, the manager itself is only a stub, since the interface
used to communicate to Chrome is still undecided.

BUG=b:179711675
TEST=- Run all unit tests.
     - Run typecd.ModeSwitch 30 times.

Change-Id: I2da477fd5de9248b650680571457a3eb3dbec942
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2694839
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>
Tested-by: Prashant Malani <pmalani@chromium.org>