patchpanel: redirect system DNS when a VPN is connected
When a VPN client is connected, shill updates the content of
/etc/resolv.conf to point to the nameservers advertised by the VPN
network. If these nameservers can only be accessed through the VPN
connection, any system services or process that is exclusively routed
through the VPN network cannot anymore resolve hostnames. This prevents
update_engine from working correctly when a VPN is connected.

This issue happened to be partially mitigated for full tunnel Android
VPNs because of another bug in shill (b/177621573) that created
additional routing rules such that system processes could typically be
routed through the Android VPN connection. b/177621573 was recently
fixed in m90 by Change-Id I75c715995aea2c70a2183afa8fd02643e40175c8.

Until the DNS proxy service that proxies all DNS traffic on the DUT is
launched (b/178991738) and eventually resolves the ambiguity caused by
a single /etc/resolv.conf file when a VPN is connected, this patch
addresses the issue by setting DNAT rules that redirect DNS queries
from local processes which are not tagged with the VPN intent fwmark
(0x8000) in mangle PREROUTING. This allows update_engine to always work
as expected when a VPN is connected.

BUG=b:178331695
TEST=Unit tests pass. Checked that ping, dig, etc can resolve hostnames
from the shell when a VPN is connected. Checked that "Check for
updadates" button in Chrome OS settings generates a DNS query on the
default physical network with the correct nameserver target.

Change-Id: I448dd633b2d78a6baab8f47c3278e516114bee95
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2692453
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>