installer: Use temp dir to migrate extensions.
Currently, the script creates a destination directory,
and import extensions directly.
In this approach, the directory is temporarily owned by root
and shares some intermediate state during the importing.

This CL starts to use .tmp suffixed directory for the import,
and after everything is done, rename to the proper name.

BUG=chromium:1132384
TEST=Remove the directory, and reboot.

Change-Id: I2b610cd76b0f2358f2c2fcda7491e35cc85ea098
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2696308
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>