shill: Fix all to CellularCapability3gpp::SetupLocation
Location polling was added here:
https://chromium-review.googlesource.com/c/aosp/platform/system/connectivity/shill/+/412192/17/cellular/cellular_capability_universal.cc#294

There is no comment in the code or the CL suggesting why it would
only get called inside an error.IsFailure() clause. It is very
likely a mistake.

Since IsLocationUpdateSupported() is only true if:
cellular()->mm_plugin() == kTelitMMPlugin

We might want to remove the code entirely, but for now this at
least moves the code to the logical location.

BUG=none
TEST=unit tests

Change-Id: I2714d53679ee02cfd5a455650346281beb1e847c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2698868
Tested-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Pavan Holla <pholla@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>