shill: CellularCapability3gpp: Remove deferred enable
The deferred_enable_modem_callback_ logic broke when the
lifetime of Cellular and Modem was detached.

This fixes an issue where Cellular was never getting
enabled if the Modem started in a "wrong" state, e.g.
when no SIM is in the active slot.

This also fixes an issue where Cellular would become
disabled whenever the Modem becaame disabled, which is
no longer desired.

BUG=b:179917598
TEST=unit tests, tast tests

Change-Id: I03505c93815be80641f714261e64e52f66d8b8f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2698879
Tested-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Pavan Holla <pholla@google.com>
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>