dns-proxy: IPConfig changes applied on property change
For the system and ARC proxies, anytime the DNS configuration of the
tracked device changes, it needs to be propagated to the resolver. The
default proxy (by definition) only uses the default device so it can be
safely skipped on this property change signal. The system proxy needs to
participate here since it can be tracking the physical device underlying
the active VPN (which would be the default device).

BUG=b:171157837
TEST=units
TEST=flashed DUT, started VPN, manually changed custom nameservers for
wifi network, observed change propagated to system proxy and resolver.

Change-Id: I0065bd53475fc7c185ad710f66e0e5497baa2dfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2699050
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Tested-by: Garrick Evans <garrick@chromium.org>