shill: modify the raw signal strength to UI signal strength conversion logic
The current implementation of WiFiService::SignalToStrength considers
the raw wifi signal strength as linear but in reality the raw signal
strength and its relation with signal quality is not linear. This
patches changes this logic and considers the raw signal strength as
non-linear to convert to UI signal strength.
Modify the unit test to validate that the categorization logic works as
expected.
The signal strength can vary from -20 dBm(Perfect) to -90 dBm(Poor).
The UI maps signal strength [0-100] and corresponding mathematical equation
to convert signal strength in dBm to UI signal strength[100-0] is as:
[100-75] -> 4 Bars in UI y=x+125 for Signal Strength -20 to -50 dBm
[75-50]  -> 3 Bars in UI y=2.5x+200 for Signal Strength -50 to -60 dBm
[50-25]  -> 2 Bars in UI y=2.5x+200 for Signal Strength -60 to -70 dBm
[25-0]   -> 1 Bar in UI y=1.25x+112.5 for Signal Strength -70 to -90 dBm

BUG=b:170208961
TEST=The conversion of raw signal strength to UI signal strength[0-100]
matches the added equations and the UI response looks correct.

Change-Id: If08bee51015fd0fa58ac0a6828cd46b50feeba1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2703231
Tested-by: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Abhishek Kumar <kuabhs@chromium.org>