dns-proxy: Restart default proxy for VPN if needed
If the dns-proxy service restarts while the default proxy is running
(e.g. because a VPN is on), it needs to be able to restart the default
proxy at this time. This patch uses shill's OnDefaultDeviceChanged
callback rather than the OnDefaultServiceChanged callback since the
latter is not triggered when the client is initialized, whereas the
former does.

Verified with both l2tp and openvpn.

BUG=b:174802586
TEST=flashed DUT, started VPN, restarted service, verified default proxy
child process restarts correctly.

Change-Id: I67a5c3fecce7278d7de7d1dbb26ebb6692e5c49f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2702994
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Tested-by: Garrick Evans <garrick@chromium.org>