power: Add ambient light sensor watcher
Adds an AmbientLightSensorWatcher class for detecting when external
ambient light sensors are connected/disconnected. The interfaces are
very similar to the existing DisplayWatcher class. It uses udev to
detect when IIO devices are connected/disconnected. The
AmbientLightSensorWatcher is only enabled if the
external_ambient_light_sensor pref flag is set.

BUG=b:177891460
TEST=unit tests

Change-Id: Iee7dc0cf2779472c48c1e59d185f4358c3cdc373
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2725231
Tested-by: Pablo Ceballos <pceballos@chromium.org>
Commit-Queue: Pablo Ceballos <pceballos@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>