chromeos-config: Add wedge-reboot-delay-ms field to modem
This new parameter will allow to tune the delay after which we pulse
the cellular modem reset GPIO if it hasn't appeared on the USB bus.

This will be used as a workaround on Hatch where the modem has an issue
on warm reboot.

BUG=b:180507051
TEST=unit-tests

Change-Id: I5c8cc4eb0b4ce89620bacf48ce33452cf5575b1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2703257
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>