modemfwd: allow to tune the delay used to reboot wedged modem
if the cros-config defines the new modem/wedge-reboot-delay-ms attribute,
uses it instead of the 5-minute default delay to decide when to check
whether the modem has appeared on the USB bus and reset it through its
reboot GPIO if it is wedged.

BUG=b:180507051
TEST=unit-tests
TEST=on Jinlon, do the reboot process described in b/179795020#comment3
and see the LTE modem coming back less than 25 second after boot and
attaching to the LTE network.

Change-Id: I6a2b854dcf06d1a08db5f4f587af6af13f4d2faf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2703258
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>