arc: arcvm-mount-myfiles: Stop setting up MyFiles in concierge namespace
Now we are going to revert the change to start ARCVM on login screen:
crrev.com/c/2706539. With the revert, the MyFiles directory is already
set up inside the concierge namespace when the arcvm-mount-myfiles
Upstart job starts, and nothing should be done by the script.

Note that the said revert is expected to be a temporary one. We would
need to set up the MyFiles directory again once the change is relanded.

BUG=b:178353348
BUG=b:178655981
TEST=Confirm that MyFiles works in ARCVM with the revert CL.

Change-Id: Ia2798e1b86037006b7832350462746b1850882d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2704741
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>