arc: arcvm-forward-pstore deletes destination file if it exists
This change modifies arcvm-forward-pstore so that it deletes the
destination pstore file, if it exists, before forwarding the contents.
This prevents permission errors from brillo::SafeFD::MakeFile().

This change also modifies ForwardContents() so that it returns whether
or not it was successful. If ForwardContents() fails, the timer will be
stopped so it will not be retried. This is to prevent log spam.

BUG=b:179615963
TEST=Change pstore file in /home/root/<hash>/crosvm to unexpected mode.
 Log out, then log in. Check that pstore file is recreated with correct
 mode.

Change-Id: I5ac6bdc5049c0e5cc52d7d4035f0b894328e3fb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2713192
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Commit-Queue: Willie Koomson <wvk@google.com>
Tested-by: Willie Koomson <wvk@google.com>