sepolicy: Use arcvm switch to remove arc_cts_fails_release for arcvm
Some of the cryptohomed and cryptohome_namespace_mounter allow rules
were added with the macro arc_cts_fails_release which releases the
failure of the CTS test but also convert the domain to permissive.
Using the arcvm switches, remove this macro for arcvm devices.

BUG=b:180766854
TEST=Make the policy deficient and emerge/deploy to the device.
See that Cryptohome.* tests fail and also see the selinux violation for
the missing allow rule in the audit.log file.

Change-Id: I06e4e2b50d370dd004322d074f0d18be52ec7a57
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2714048
Commit-Queue: Betul Soysal <betuls@google.com>
Tested-by: Betul Soysal <betuls@google.com>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>