lorgnette: Make timeout dynamic
In order to avoid consuming resources when scanning isn't in use,
lorgnette exits after an idle timeout.  However, this means that the
list of available scanners doesn't update during that timeout period.
If the user keeps hitting retry in the UI, lorgnette will never hit the
idle timeout, so new scanners will never be detected.

In the long run, we need to get better at responding to USB changes
dynamically.  In the short run, make the default timeout much shorter so
that the problem isn't as easy to trigger.  However, if lorgnette exits
in the middle of a scan, it will lose its internal state.  Combine the
two ideas by splitting the idle timeout into a short timeout used for
ListScanners() and GetCapabilities(), plus a long timeout used during
scanning operations.

BUG=b:179717254
TEST=lorgnette_cli

Change-Id: Ic7058cbd59662efc6ea7f39e043759a97fa9d0d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2707625
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Jesse Schettler <jschettler@chromium.org>