power: External ALS to display matching and handler
Adds logic for matching external ambient light sensors with external
displays in the ExternalBacklightController. The algorithm works by
calculating an association score between each ALS and each display based
on how similar their syspaths are.

Adds an ExternalAmbientLightHandler class to act as a shim between the
ExternalBacklightController and an AmbientLightHandler. The
ExternalAmbientLightHandler owns the AmbientLightHandler and
corresponding AmbientLightSensorInterface. It also stores the display's
path to be passed to the ExternalBacklightController so that the correct
display's brightness can be adjusted.

BUG=b:177891460
TEST=Unit tests

Change-Id: I141e254250904fd2f840fc1e6991fcf1c79323e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2725235
Tested-by: Pablo Ceballos <pceballos@chromium.org>
Commit-Queue: Pablo Ceballos <pceballos@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>