Add os_install_service
This is a DBus system service used to install the OS to disk. It
exposes a single method, `StartOsInstall`. This method takes no
parameters; the service chooses an appropriate disk to install to
without any user input. Updates are provided with the
`OsInstallStatusChanged` signal. Currently there is no
percentage-complete report, the signal just indicates if the install
succeeded or failed. The signal also includes the install log so that
error details can be presented.

This service (when included in the OS image) only runs when the OS is
live booted from an installer image. This is checked in the upstart
script (conf/os_install.conf) by running `is_running_from_installer`,
which compares the sizes of the root-A and root-B partitions. If they
are the same size, then the OS is considered installed, whereas if the
sizes are different then the OS is running from an installer image
with a stub root-B partition.

BUG=b:182386612
TEST=cros_workon_make --test os_install_service
TEST=USE="$USE os_install_service" ./build_packages
     ./build_image test
     Run it in a VM with `-hdb` pointed at an empty disk image.
     Start dbus-monitor to watch signals:
       dbus-monitor --system sender=org.chromium.OsInstallService
     Start the install process:
       sudo -u chronos dbus-send --print-reply --system \
        --dest=org.chromium.OsInstallService \
        /org/chromium/OsInstallService \
        org.chromium.OsInstallService.StartOsInstall
     Wait for the monitor to show the `OsInstallStatusChanged` signal,
     which should show two strings: "Succeeded" for status, and the
     install log.
     Power off the VM and start a new VM from the installed disk image.
     Verify it boots.
     Run `initctl list | grep os_install` and verify that the install
     service isn't running.

Cq-Depend: chromium:2787390
Change-Id: I595716b840670b3207c9220ae80ff934507560b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2787389
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>