shill: simplify PortalDetector start and stop trial
This is a reland of crrev/c/2760640 Change-Id
I4c4b4734324370669bfb0debb32b52e7b592419e. In this new version,
CleanupTrial() is called before invoking |portal_result_callback_| and
the ordering is unchanged.

This patch folds together multiple functions calling into each other
when a portal detection trial is started or stopped. The result is less
code and a logical flow that is easier to follow.

Also, this patch ensures that a metric event is recorded if the http
probe fails to start. Previously no event would be recorded in that
case.

BUG=b:182745554
TEST=unit tests. Checked portal validation on WiFi and Ethernet.

Change-Id: I0caf600eb82dd97544d3b82b1d558b40d12d8a23
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2790193
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>