shill: simplify PortalDetector start and stop trial
This patch folds together multiple functions calling into each other
when a portal detection trial is started or stopped. The result is less
code and a logical flow that is easier to follow.

Also, this patch ensures that a metric event is recorded if the http
probe fails to start. Previously no event would be recorded in that
case.

BUG=b:182745554
TEST=unit tests. Checked portal validation on WiFi and Ethernet.

Change-Id: I4c4b4734324370669bfb0debb32b52e7b592419e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2760640
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>