avtest_label_detect: Add protected video capabilities
This detects the ability of the video device to be able to do decoding
of Widevine protected content over various codecs, encryption schemes
(CTR vs. CBC) and also encryption mode (CENCv1 vs. CENCv3).  CENCv1 is
only supported for H.264 in Chrome, so we only check that codec for
CENCv1.  We also skip VP8 since we don't implement that for protected
content in Chrome.

BUG=b:155501161
TEST=avtest_label_detect returns protected labels on volteer

Change-Id: If89a91460e53e48149d70efa797d363bee708f76
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2728301
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>