typecd: Add check for Intel's SVID on TBT3 Cable Identity Check
A check is necessary in the event that we encounter an active cable.
If said active cable does not have the 0x8087 object, it's not TBT3 compatible
and we must bail TBT3 Discovery.

BUG=b:181426830
TEST=Unittests pass.
Check that when using the Cable Matters 201034-BLK-3m cable, and a TBT4 dock,
typecd enters DP mode, not TBT3.
Check that when a TBT3 active cable is used (eg. Belkin F2CD085bt2M-BLK), and
a TBT4 dock, TBT3 mode is entered when screen is unlocked.

Change-Id: I63d7147df0345582a15fa36fd7a2d5c28abfbedf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2728314
Commit-Queue: Benson Leung <bleung@google.com>
Tested-by: Benson Leung <bleung@google.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>