imageloader: Always create the imageloader components path
The components path can be created whenever imageloader is started. Even
when imageloader init conf is called, it should create the components
path if it's missing.

This move is better because it doesn't require imageloader to be
started/stopped for the imageloader components path to exist. In the TLS
implementation to provision Lacros onto the device, this path can be
created by imageloader during startup instead of doing so from TLS
implementation.

BUG=b:177259613
TEST=FEATURES=test emerge-$B imageloader

Change-Id: I160d2d2ac8d414286a704e926e7ab92d36ff99de
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2737556
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Auto-Submit: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>