u2fd: Fix bug that returned assertion was using first allowed credential
When there are both U2F credentials and platform credentials in the
allowed credential list, GetAssertion will always prefer platform
credentials. The old code writes the first allowed credential id
to the assertion object in GetAssertion response, and that is wrong
because the first allowed credential will not be the asserted credential
if it's a U2F credential (lower priority).

BUG=b:182218069
TEST=Register a U2F credential followed by a platform credential (so
     that the allowed credential list comes in this order). GetAssertion
     asks for fingerprint auth and succeeds on website.

Change-Id: Iec8f13666e35fcdd41e86e5d483f11cc343bc68e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2743689
Tested-by: Yicheng Li <yichengli@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Reviewed-by: Yicheng Li <yichengli@chromium.org>
Commit-Queue: Yicheng Li <yichengli@chromium.org>