debugd: get form factor from cros_cofig instead of /etc/lsb-release
Previously, the device type/form factor was stored in /etc/lsb-release,
however this did not allow the value to be set individually for
unibuild devices. Instead, cros_config should be used to retrieve the
correct form factor on all devives. The DEVICETYPE value is still
present in in /etc/lsb-release, and cannot currently be removed, so it
needs to be filtered out of the logs.

BUG=b:173179861
TEST=viewed logs in chrome://system, DEVICETYPE was replaced
     with value from cros_config

Cq-Depend: chromium:2784367
Change-Id: I900b04dfa8061ff46a50bfa029dd9a0cbbb30e6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2744956
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Nikolai Artemiev <nartemiev@google.com>
Commit-Queue: Nikolai Artemiev <nartemiev@google.com>