debugd: Add lspci_verbose to feedback report.
The output of 'lspci -vvv' gives information about the status of the
PCIe link, for example the number of lanes and the bitrate at which the
PCIe device trained.
This is useful to notice problems with the PCIe link, which can point to
coreboot issues, SI issues, etc.
On kohaku (hatch), the output of `lspci` increases from ~1KB to ~25KB
before compression. After compression, the size increase is 0.5KB to
4.5KB.

Introduced two vectors kCommandLogsVerbose and kCommandLogsShort to
separate the verbose version of the commands that need to be used in
feedback reports from the non_verbose commands that are used in
chrome://system.

BUG=b:168563469
BUG=chromium:1126772
BUG=b:161736733
TEST=emerge-hatch debugd
Checked the chrome://system it shows "lspci"
Ran the test feedback.SysInfoPII and printed the output it only shows
"lspci_verbose".

Change-Id: I0cf964e37e467798bb961edfa5dd862291cc4644
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2746355
Tested-by: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Arowa Suliman <arowa@chromium.org>