shill: update pppoe plugin name
In crrev.com/c/2696947, rp-pppoe.so plugin was replaced with pppoe.so,
but shill is still using rp-pppoe.so plugin. Hence pppd couldn't load
rp-pppoe.so and PPP connection is failing.

2021-03-09T07:25:33.853168Z ERR pppd[6803]: Couldn't load plugin rp-pppoe.so
2021-03-09T07:25:33.854962Z INFO shill[2531]: INFO shill: [service.cc(384)] Disconnecting from ppoe_2: OnPPPDied

Change rp-pppoe.so to pppoe.so in ppp_daemon.cc

BUG=b:182282096
TEST=Compiled and verified pppd could load the plugin and PPPoE
connection is successful.

Change-Id: Ibfb95ccf065d9574f27139cd5be67657a3b7fe51
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2746456
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Raju Konduru <rkonduru@google.com>
Commit-Queue: Raju Konduru <rkonduru@google.com>