imageloader: Increase polling and dmsetup timeouts
Imageloader will not become a single process to handle both parsing
and mounting. So in order to account for future growth of mount times,
increase the timeouts for polling and dmsetup calls that the main
process waits on for the helper to finish.

BUG=b:172220337
TEST=FEATURES=test emerge-$B imageloader

Change-Id: I3224a10f90b72238888b199943564f863bbf9519
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2746801
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>