cryptohome: Use correct LE Credentials check in ResetLECredential.
This fixes a crash by using the same IsLECredentials() check that the
calling code does. Note that there is confusingly two places where we
identify a credential as LE - in the flags and in the policy. This also
makes the CHECK() an error, as a corrupted SerializedVaultKeyset should
not cause a crash.

BUG=b:182345624
TEST=tast run ${DUT_IP} hwsec.*

Change-Id: I44567d10cbd0b6fbf4a6c432ebc8c7c5f2b29d6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2750374
Tested-by: Greg Kerr <kerrnel@chromium.org>
Auto-Submit: Greg Kerr <kerrnel@chromium.org>
Commit-Queue: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>