bootstat: Convert to C++
Basic conversion to C++, the core of the functions still use raw
C APIs, and the tests are not updated yet. This structure should
allow us to mock BootStat in the next CL.

The logic in `append_logdata` is split in OpenEventFile and
CopyFromFile as we plan to have data that is _not_ coming from
simple files in the future.

BUG=b:182417100
TEST=FEATURES=test emerge-amd64-generic --nodeps -v bootstat
TEST=USE=-cros-debug emerge-asurada -v bootstat && \
     cros deploy $IP bootstat && \
     ssh $IP "rm -f /tmp/*-x; bootstat x; grep . /tmp/*-x"

Change-Id: I972ab9f82997f22587dae59b44d475b1d8230637
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2750853
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>