bootstat: Start converting unit tests
Move all the functions that need to interact with the system to
BootStatSystem class, and inherit that in the tests.
Currently, only one function needs to be overridden, but there
will be more in the future.

Also, don't try to use the .so library for unit tests, as we'd
need to export too many functions.

BUG=b:182417100
TEST=cros_workon_make --board=amd64-generic bootstat --test

Change-Id: Ibad4b1fac8a5ed8bad00c6bd278bf7c17782a0a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2752690
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>