runtime_probe: Implement a variation for factory environment
In factory, the ability to change to a different version of binaries
without modifying the rootfs is the key requirement.  On the other
hand, security is not that critical in such scenario.  Hence, this CL
introduces a different build target that generates a variation of
the runtime_probe binary that runs out-of-the-box, i.e. it doesn't
rely on `debugd` to invoke helper sub-processes.

In addition, this CL removes the assumption that the format string
`%lx` for printf maps to 64-bit integer by using `PRIx64` from
`inttypes.h` instead.

BUG=b:182536856
TEST=manual test

Change-Id: Iad279097f44685eddc296dffcf5d1065e6df8ad0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2755373
Reviewed-by: Kevin Lin <kevinptt@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Yong Hong <yhong@chromium.org>
Commit-Queue: Yong Hong <yhong@chromium.org>