cryptohome: lockbox-cache: Remove bootstat calls
It does not appear that anybody is consuming that data, so let's
remove them from init scripts.

The overhead is small (~2ms per call), but there is still no reason
to pay it (we could just link in libbootstat in lockbox-cache if
we really wanted these metrics).

BUG=none
TEST=CQ

Change-Id: I2c633c7213d27ec21def7711dc8c06838af77d7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2755377
Reviewed-by: John L Chen <zuan@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>