shill: vpn: Support multiple peers in WireguardDriver
This patch replaces the properties for a peer with a single property
"Wireguard.Peers". This property is stored as an array of json-encoded
strings in storage, and can be read/write via RPC as a Stringmaps (i.e.,
vector<map<string, string>>). Inside WireguardDriver, we use a
Stringmaps member |peers_| to hold this property, and override several
functions from VPNDriver to provide logic for connecting |peers_| with
storage and PropertyStore.

Inside a peer, the "PresharedKey" needs special handling since it should
be read-only via RPC.

BUG=b:177876632
TEST=unit_tests
TEST=manually read/write properties via D-Bus

Change-Id: Ib958ce3427398276f64745a38437a4f519e957cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2758330
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>