shill: Refactor DISABLE_WAKE_ON_WIFI.
Make WiFi::wake_on_wifi_ optional to reduce most DISABLE_WAKE_ON_WIFI
usages. When DISABLE_WAKE_ON_WIFI is defined, the public methods of
WakeOnWiFi is just no-op or collecting metrics like SuspendDuration
when WoWiFi on/off, which is not very valuable if wake on WiFi is not
planned to be enabled on the board. So, we can just skip the calls in
the case.

Also, some codes in !defined(DISABLE_WAKE_ON_WIFI) depend on
NL80211_ATTR_WOWLAN_TRIGGERS attribute which is introduced in kernel
v3.0. That should be old enough that we can always compile it in shill.
We may still use some newer triggers like NL80211_WOWLAN_TRIG_NET_DETECT
(introduced in v3.19), but those can usually be resolved by syncing
nl80211.h and check NL80211_ATTR_WOWLAN_TRIGGERS_SUPPORTED before use
without much trouble. (this is what we did in v3.18 kernel).

This change still introduces a noticeable behavior change that when
DISABLE_WAKE_ON_WIFI is set, all wake on WiFi related Properties like
"WakeOnWiFiFeaturesEnabled" will not appear on the WiFi device. The
setter may need to check the existence of the properties first.

BUG=b:172208175
TEST=unit tests
TEST=run suite:wifi_matfunc tests

Change-Id: I37c9eb70ab0cf4551f310c27d6636bb31a927074
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2763546
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Tested-by: Yen-lin Lai <yenlinlai@chromium.org>
Commit-Queue: Yen-lin Lai <yenlinlai@chromium.org>