vm_tools: Remove sommelier's "base/" includes
Sommelier is not linked against libchrome, and is built in an
environment where libchrome is not available.

In crrev.com/c/2741767 a mass-refactoring incorrectly modified a
sommelier file to include a libchrome header. Whilst this may have
worked on the CQ, where libchrome headers were available, it broke
builds not on the CQ.

This CL removes that include.

BUG=b:182837372
TEST=build sommelier using meson

Change-Id: Ia7b85e69efddadae759bb2d754b404692fe12d27
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2763302
Reviewed-by: Qijiang Fan <fqj@google.com>
Tested-by: Nic Hollingum <hollingum@google.com>
Commit-Queue: David Riley <davidriley@chromium.org>