shill: Cellular: Clean up eid in CellularService
This cleans up the use of eid in CellularService, making it
an immutable property set on construction.

The sim_card_id_ member is replaced with a getter identical to
the one in Cellular (eid || iccid).

We keep the SimCardId property in the profile for now to
simplify retrieval of services associated with a SIM card.

This also ensures that secondary SIM slot properties are
updated when they change, e.g. when a pSIM hotswap occurs
while the eSIM slot is primary.

BUG=b:182606954
TEST=Tast tests + unit tests + manual testing

Change-Id: I5cb9018239d16bea5351b17b4c8420864e638415
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2767598
Commit-Queue: Steven Bennetts <stevenjb@chromium.org>
Tested-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Pavan Holla <pholla@google.com>